Move archive logic out of CLI
